--- Backend/auth.js
@@ -1,1 +1,0 @@
-// file to handle the auth routes
